% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_dis.R
\name{cond_dis}
\alias{cond_dis}
\title{Simulating cross validation splits when only group means are avaiable.}
\usage{
cond_dis(mu, sigma, tau, theta)
}
\arguments{
\item{mu}{marginal mean of group means.}

\item{sigma}{marginal covariance matrix for group means.}

\item{tau}{sum of the two cv splits.}

\item{theta}{linear constrain of the two cv splits.}
}
\value{
return a conditional mean and a covariance matrix.
}
\description{
Simulating cross validation splits when only group means are avaiable.
}
\references{
Peysakhovich, A., & Eckles, D., Learning causal effects from many randomized experiments
using regularized instrumental variables.
\emph{Proceedings of the 2018 World Wide Web Conference} (2018, April). URL
https://dl.acm.org/doi/10.1145/3178876.3186151'

Owen, Art B. \emph{Monte Carlo Theory, Methods and Examples.} (2016). URL
 http://statweb.stanford.edu/~owen/mc/
}
