% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gIVCVs.R
\name{gIVCVs}
\alias{gIVCVs}
\title{Generalized instrumental variable with cross validation using summary statistics.}
\usage{
gIVCVs(
  formula,
  id,
  ctrl_id,
  ctrl_sd_x,
  ctrl_sd_y,
  data,
  subset = NULL,
  weights = NULL,
  wt = NULL,
  instruments = NULL,
  na.action = NULL,
  L0 = F,
  theta = t(c(1, 1)),
  ...
)
}
\arguments{
\item{formula}{representing relations among variables, following the syntax used in `lm`.
`r y ~ x | z` means the first stage OLS is `r x ~ z` and the second stage is `y ~ x`.}

\item{id}{experiment id.}

\item{ctrl_id}{indicators for the control arms.}

\item{ctrl_sd_x}{estimated standard error for the group-level means for the independent variables.}

\item{ctrl_sd_y}{estimated standard error for the group-level means for the dependent variables.}

\item{data}{data frame with all the variables in the model specified in formula.}

\item{wt}{for the second stage regression.}

\item{L0}{a logic statement whether to apply the L0 regularization and cross validation or not.}
}
\value{
a list that contains the final tuning parameter, the data set after L0 regularization and a lm fitted object.
}
\description{
Generalized instrumental variable with cross validation using summary statistics.
}
\examples{
# true effect is .1, .2, .3
J = 2000
I = 100
data_gIVCVs <- sim_gIV(beta_ctrl = .3,
                      beta_X = c(.1, .1, .1),
                      J = J, I = I)
data_gIVCVs \%>\% head()
# generalized IVCV with L0
gIVCVs(y ~ x1 + x2 + x3,
      id = id,
      ctrl_id = ctrl_id,
      ctrl_sd_x = sqrt(2/J) \%>\%
        rep_col(3) \%>\%
        rep_row(J),
     ctrl_sd_y = rep(sqrt(3/J), J),
     L0 = T,
     data = data_gIVCVs)$fit
}
\references{
Peysakhovich, A., & Eckles, D., Learning causal effects from many randomized experiments
using regularized instrumental variables. \emph{Proceedings of the 2018 World Wide Web Conference} (2018, April).
URL https://dl.acm.org/doi/10.1145/3178876.3186151
}
