% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVCV.R
\name{IVCV}
\alias{IVCV}
\title{Instrumental variable with cross validation}
\usage{
IVCV(
  formula,
  id,
  data,
  subset = NULL,
  weights = NULL,
  wt = NULL,
  ctrl_id = 1,
  instruments = NULL,
  na.action = NULL,
  L0 = F,
  raw = F,
  ...
)
}
\arguments{
\item{formula}{representing relations among variables, following the syntax used in `lm`.
`y ~ x | z` means the first stage OLS is `x ~ z` and the second stage is `y ~ x`.}

\item{id}{experiment id.}

\item{ctrl_id}{indicators for the control arm.}

\item{instruments}{model for the first stage OLS, omitted when use `|` syntax in formula.}

\item{L0}{a logic statement whether to apply the L0 regularization and cross validation or not.}

\item{raw}{if run TSLS with the raw data}
}
\value{
a list that contains the final tuning parameter, the data set after L0 regularization and a lm fitted object.
}
\description{
Instrumental variable with cross validation
}
\examples{
# true effect: .1 .2 .3
I = 100
J = 500

dat_raw <- sim_IV(I = I, # experiment size
                  J = J, # number of experiment
                  beta_X = c(.1, .2, .3), # effects
                  ctrl_id = 1) # the indicator of treatment group
head(dat_raw)
# TSLS
fit_raw_tsls <- IVCV(y ~ x1 + x2 + x3 | factor(id),
                    id = id,
                    data = dat_raw,
                    L0 = F)$fit
# IVCV
fit_raw_ivcv <- IVCV(y ~ x1 + x2 + x3 | factor(id),
                    id = id,
                    ctrl_id = 1,
                    data = dat_raw,
                    L0 = T)$fit


}
\references{
Peysakhovich, A., & Eckles, D., Learning causal effects from many randomized experiments
using regularized instrumental variables.
\emph{Proceedings of the 2018 World Wide Web Conference} (2018, April). URL
https://dl.acm.org/doi/10.1145/3178876.3186151'
}
\keyword{instrumental}
\keyword{variable}
