% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cv.R
\name{make_cv}
\alias{make_cv}
\title{Make cv split indicators within tibble.}
\usage{
make_cv(data, group_id, n_cv = 2)
}
\arguments{
\item{data}{tibble object.}

\item{group_id}{number of cv splits.}
}
\value{
a tibble with id numbers for cv splits.
}
\description{
Make cv split indicators within tibble.
}
\examples{
tibble(a =1:10, exp_id = rep(1:5, 2)) \%>\%
make_cv(group_id = exp_id, 2)
}
