% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dummy.R
\name{make_dummy}
\alias{make_dummy}
\title{Helper function - make dummy matrices.}
\usage{
make_dummy(fac_var)
}
\arguments{
\item{fac_var}{take a vector and turn into a dummy coded matrix.}
}
\description{
Helper function - make dummy matrices.
}
