% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange.R
\name{rearrange}
\alias{rearrange}
\title{Helper function - reshape a list with atomic elements.}
\usage{
rearrange(list, n_list = 2, n_var = 3)
}
\arguments{
\item{list}{a list with atomic elements.}

\item{n_list}{number of nested lists.}

\item{n_var}{number of atomic elements.}
}
\value{
a reshaped list.
}
\description{
Helper function - reshape a list with atomic elements.
}
