% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_RE.R
\name{select}
\alias{select}
\title{Shiny app for the data generating process of random intercept models.}
\usage{
select(.data, ...)
}
\value{
return a new Shinyapp session.
}
\description{
Shiny app for the data generating process of random intercept models.
}
