% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gIV.R
\name{sim_gIV}
\alias{sim_gIV}
\title{Simulating data for IV with multitple experiments.}
\usage{
sim_gIV(
  mu_Z = c(0, 0, 0),
  beta_ctrl = 0.2,
  summary_data = T,
  sig_Z = matrix(c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1), nrow = 3),
  mu_U = c(0, 0, 0),
  sig_U = matrix(c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1), nrow = 3),
  beta_X = c(0.1, 0.2, 0.3),
  beta_UX = matrix(c(0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5), nrow = 3),
  beta_UY = c(1, 2, 3),
  mu_e_X = c(0, 0, 0),
  sig_e_X = matrix(c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1), nrow = 3),
  mu_e_Y = 0,
  sig_e_Y = 1,
  J = 50,
  I = 30
)
}
\arguments{
\item{mu_Z}{expectation for the true effects.}

\item{beta_ctrl}{true effect of the experiment arm.}

\item{summary_data}{return a data, averaged across experiments and arms.}

\item{mu_U}{expectation of the multidimensional confounders.}

\item{sig_U}{covariance matrix of the multidimensional confounders.}

\item{beta_X}{true effects of the causal variables of intrest.}

\item{beta_UX}{regression coefficients of the confounders on the causal variables of interest.}

\item{beta_UY}{regression coefficients of the confounders on the outcome variable.}

\item{mu_e_X}{expectation of the first stage errors - within group errors.}

\item{sig_e_X}{covariance matrix of the first stage errors - within group errors.}

\item{mu_e_Y}{expectation of the second stage errors - between group errors.}

\item{sig_e_Y}{variance of the second stage errors - between group errors.}

\item{J}{number of experiments.}

\item{I}{size of each experiment arm.}
}
\description{
Simulating data for IV with multitple experiments.
}
\examples{
# true effect is .1, .2, .3
J = 2000
I = 100
sim_gIV(beta_ctrl = .3,
        beta_X = c(.1, .1, .1),
        J = J, I = I)
}
