% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sIVCV.R
\name{sIVCV}
\alias{sIVCV}
\title{Instrumental variable with cross validation with summary statistics.}
\usage{
sIVCV(
  formula,
  id,
  data,
  ctrl_sd_x,
  ctrl_sd_y,
  subset = NULL,
  weights = NULL,
  wt = NULL,
  ctrl_id = 1,
  instruments = NULL,
  na.action = NULL,
  L0 = F,
  ...
)
}
\arguments{
\item{formula}{representing relations among variables, following the syntax used in `lm`.
`y ~ x | z` means the first stage OLS is `x ~ z` and the second stage is `y ~ x`.}

\item{id}{experiment id.}

\item{wt}{for the second stage regression.}

\item{ctrl_id}{indicators for the control arm.}

\item{instruments}{model for the first stage OLS, omitted when use `|` syntax in formula.}

\item{L0}{a logic statement whether to apply the L0 regularization and cross validation or not.}

\item{raw}{if run TSLS with the raw data}
}
\value{
a list that contains the final tuning parameter, the data set after L0 regularization and a lm fitted object.
}
\description{
Instrumental variable with cross validation with summary statistics.
}
\references{
Peysakhovich, A., & Eckles, D., Learning causal effects from many randomized experiments
using regularized instrumental variables.
\emph{Proceedings of the 2018 World Wide Web Conference} (2018, April). URL
https://dl.acm.org/doi/10.1145/3178876.3186151'
}
\keyword{instrumental}
\keyword{variable}
