% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sIVCV.R
\name{rearrange}
\alias{rearrange}
\title{A test Function}
\usage{
rearrange(list, n_list = 2, n_var = 3)
}
\arguments{
\item{agree}{Testing if the package works well.}
}
\description{
A test Function
}
\keyword{instrumental}
\keyword{variable}
